+++ src/web/assets/AuthorToolbarAssets.php
@@ -5,7 +5,6 @@ namespace digitalastronaut\craftauthortoolbar\web\assets;
 use craft\web\AssetBundle;
 
 class AuthorToolbarAssets extends AssetBundle {
-    
     public function init(): void {
         $this->sourcePath = "@digitalastronaut/craftauthortoolbar/web/assets/dist";